stock_picking_cancel: fix getting quant
When using package during the delevery process.
The quant in the location_dest_id will have a package
but not the quant in the location_id.
The domain is wrong and the quant in the location_id is not found.
We have to filter on the result_package_id in the location_dest_id
and on the package_id for the location_id